+++ src/helper.ts
@@ -1,11 +1,25 @@
 import joplin from "api";
 import * as path from "path";
-import { exec } from "child_process";
+import { execFile } from "child_process";
 import { promisify } from "util";
 import { copyFileSync } from "fs-extra";
+import { copySync } from "fs-extra";
 import { moveSync } from "fs-extra";
+import * as fs from "fs-extra";
 
 export namespace helper {
+  export async function getPluginVersion(): Promise<string> {
+    const installationDir = await joplin.plugins.installationDir();
+    try {
+      const manifest = JSON.parse(
+        fs.readFileSync(path.join(installationDir, "manifest.json"), "utf8")
+      );
+      return manifest.version;
+    } catch (error) {
+      return "n/a";
+    }
+  }
+
   export async function validFileName(fileName: string) {
     var regChar = /[:*?"<>\/|\\]+/; // forbidden characters \ / : * ? " < > |
     var rexNames = /^(nul|prn|con|lpt[0-9]|com[0-9])(\.|$)/i; // forbidden file names
@@ -103,10 +117,27 @@ export namespace helper {
     fsWorkaroundLinux: boolean
   ): Promise<boolean> {
     if (process.platform == "linux" && fsWorkaroundLinux === true) {
-      var execPromise = promisify(exec);
-      await execPromise(`cp -r '${src}' '${dst}'`);
+      var execPromise = promisify(execFile);
+      await execPromise("cp", ["-r", src, dst]);
     } else {
       copyFileSync(src, dst);
+    }
+
+    return true;
+  }
+
+  // Workaround for "ENOTSUP: operation not supported on socket" #98
+  // https://github.com/JackGruber/joplin-plugin-backup/issues/98
+  export async function WorkaroundCopyFolder(
+    src: string,
+    dst: string,
+    fsWorkaroundLinux: boolean
+  ): Promise<boolean> {
+    if (process.platform == "linux" && fsWorkaroundLinux === true) {
+      var execPromise = promisify(execFile);
+      await execPromise("cp", ["-r", src, dst]);
+    } else {
+      copySync(src, dst);
     }
 
     return true;
@@ -121,8 +152,8 @@ export namespace helper {
     overwrite: boolean = false
   ): Promise<boolean> {
     if (process.platform == "linux" && fsWorkaroundLinux === true) {
-      var execPromise = promisify(exec);
-      await execPromise(`mv '${src}' '${dst}'`);
+      var execPromise = promisify(execFile);
+      await execPromise("mv", [src, dst]);
     } else {
       moveSync(src, dst, {
         overwrite: overwrite,