+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## not released
 
+## v1.5.1 (2025-12-31)
+
+- Fix: Joplin versions without multiple instances support
+- Fix: Backup of plugins #99
+- Fix: Security issue on escaping paths for copy workaround
+
 ## v1.5.0 (2025-12-28)
 
 - Fix: Add workaround for "ENOTSUP: operation not supported on socket" #98